[Client] Added human-oriented messages to some actions

+++ StartFolio/ClientApp/app/components/gallery/gallery.component.ts
@@ -54,6 +54,12 @@ export class GalleryComponent implements OnInit {
 
     deletePage()
     {
+        let willDelete = confirm("Do you really want to delete this component?");
+
+        if (!willDelete) {
+            return;
+        }
+
         this.pageService.deletePage(this.position);
     }
 }
