+++ StartFolio/ClientApp/app/components/person/person.component.ts
@@ -52,6 +52,12 @@ export class PersonComponent implements OnInit {
     }
 
     deletePage() {
+        let willDelete = confirm("Do you really want to delete this component?");
+
+        if (!willDelete) {
+            return;
+        }
+
         this.pageService.deletePage(this.position);
     }
