+++ StartFolio/ClientApp/app/components/team/team.component.ts
@@ -59,6 +59,12 @@ export class TeamComponent implements OnInit {
     }
 
     deletePage() {
+        let willDelete = confirm("Do you really want to delete this component?");
+
+        if (!willDelete) {
+            return;
+        }
+
         this.pageService.deletePage(this.position);
     }
